fix: view.tsx 분리
