+++ src/pages/PasswordConfirm/PasswordConfirm.tsx
@@ -1,64 +1,15 @@
-import React, { useState } from "react";
-import InputText from "components/common/inputText";
-import InputForm from "components/common/InputForm";
+import PasswordConfirmView from './PasswordConfirm.view'
 
-import { useNavigate } from 'react-router-dom';
-
-import './passwordConfirm.css';
-
-export interface PwdConfirmProps {
-  isFailed?: boolean;
-}
-
-const PasswordConfirm = ({
-  isFailed = false
-}: PwdConfirmProps) => {
-  
-  const navigate = useNavigate();
-
-  const [modeValue, setModeValue] = useState<"normal" | "focus" | "warning">(isFailed ? "warning" : "normal")
-  const [stateValues, setStateValues] = useState({
-    currentPwd: ""
-  });
-
-  const clickSubmit = () => {
+const PasswordConfirm = () => {
+  const clickSubmit = (data: {currentPwd: string}) => {
     console.log("현재 비밀번호 확인 테스트");
-    console.log("currentPwd : ", stateValues.currentPwd);
+    console.log("currentPwd : ", data.currentPwd);
 
     //TODO: 비밀번호 확인 API 연동 로직 구현
 
   }
 
-  return (
-    <InputForm 
-      showBackBtn={true} 
-      onClickBackBtn={true}
-      submitButtonText={"확인"}
-      disabledSubmitButton={true}
-      onSubmit={clickSubmit}
-    >
-      <div className='pwdConfirm-wrap'>
-        <div
-          className={['div-info', (modeValue != "normal")? 'div-hidden' : ''].join(' ')}
-        >
-          <p>본인 확인을 위해</p>
-          <p>비밀번호를 입력해주세요</p>
-        </div>
-        <div className='input-comp-wrap' onFocus={() => setModeValue('focus')} onBlur={() => setModeValue('normal')}>
-          <InputText
-            type='password'
-            inputId="confirmPwd"
-            mode={modeValue}
-            widthSize="medium"
-            placeholder='현재 비밀번호'
-            onChangeValue={(value: string) =>
-              setStateValues({ currentPwd: value})
-            }
-          />
-        </div>
-      </div>
-    </InputForm>
-  );
+  return <PasswordConfirmView isFailed={false} clickSubmit={clickSubmit}/>
 };
 
 export default PasswordConfirm;
