+++ src/pages/SignIn/SignIn.tsx
@@ -1,8 +1,16 @@
 import SignInView from "./SignIn.view";
 
 const SignIn = () => {
+
+  const clickSubmit = (data: {email: string, password: string}) => {
+    //TODO: 클릭 이벤트 적용하기
+    console.log("로그인 테스트");
+    console.log("email : ", data.email);
+    console.log("password: ", data.password);
+  }
   
-  return <SignInView isLoggedIn ={true} />
+  //TODO: isLoggedIn 값 변경 필요
+  return <SignInView isLoggedIn ={true} clickSubmit={clickSubmit}/>
 };
 
 export default SignIn;
